Stage 0 Module 3 Task 6 rework.

+++ src/main/java/lang/print/gaps/task6/FloatTypeCasting.java
@@ -2,6 +2,6 @@ package lang.print.gaps.task6;
 
 public class FloatTypeCasting {
     public void roundNumber(float numberToBeRounded) {
-    			System.out.println((double) Math.round(numberToBeRounded + 0.1));
+    			System.out.println (Math.round(numberToBeRounded + 0.1));
     }
 }
